requirements.txt
alembic==1.7.4
amqp==5.0.6
APScheduler==3.8.1
asgiref==3.2.10
astroid==2.4.2
automium==0.2.6
automium-web==0.1.1
autopep8==1.5.7
backports.zoneinfo==0.2.1
bcrypt==3.2.0
billiard==3.6.4.0
blinker==1.4
cached-property==1.5.2
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.26.0
colorama==0.4.4
configparser==5.1.0
cryptography==3.4.7
cycler==0.10.0
decorators==2.0.7
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.2
django-allauth==0.45.0
django-anymail==8.4
django-celery-beat==2.2.1
django-crispy-forms==1.12.0
django-filter==2.4.0
django-import-export==2.4.0
django-mathfilters==1.0.0
django-microsoft-authentication==0.1.2
django-minio-storage==0.3.10
django-phonenumber-field==5.2.0
django-rest-passwordreset==1.2.1
django-sms==0.5.0
django-timezone-field==4.2.1
django-widget-tweaks==1.4.8
djangorestframework
djangorestframework-jwt
dnspython==2.1.0
email-validator==1.1.3
et-xmlfile==1.1.0
Flask==2.0.0
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.2
gunicorn==20.0.4
idna==3.2
importlib-metadata==4.6.0
importlib-resources==5.3.0
instagram-private-api==1.6.0.0
isort==5.6.4
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
kombu==5.1.0
lazy-object-proxy==1.4.3
legacy==0.1.6
Mako==1.1.5
MarkupPy==1.14
MarkupSafe==2.0.1
matplotlib==3.3.3
mccabe==0.6.1
minio==6.0.2
msal==1.13.0
numpy==1.19.3
oauthlib==3.1.1
odfpy==1.4.1
opencv-contrib-python==4.5.2.54
openpyxl==3.0.9
pandas==1.1.5
phonenumbers==8.12.31
Pillow==8.0.1
prompt-toolkit==3.0.19
psycopg2==2.8.6
pycodestyle==2.7.0
pycparser==2.20
PyJWT
pylint==2.6.0
pyotp==2.6.0
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.19.1
python3-openid==3.2.0
pytz==2020.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.15.0
SQLAlchemy==1.4.26
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
twilio==7.2.0
typed-ast==1.4.1
typing-extensions==3.10.0.0
tzdata==2021.4
tzlocal==4.0.1
urllib3==1.26.6
views==0.3
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.1
whitenoise==5.2.0
wrapt==1.12.1
WTForms==2.3.3
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.1
